backend/pyproject.toml
[tool.poetry]
name = "intellikit"
version = "0.1.0"
description = "For doing stuff"
authors = ["Trav <travis.vas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.112.1"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
pyjwt = {extras = ["crypto"], version = "2.8.0"}
pandas = "^2.2.2"
chardet = "^5.2.0"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"
ecdsa = "^0.19.0"
ldap3 = "^2.9.1"
openai = "^1.41.0"
email-validator = "^2.1.0"
argon2-cffi = "*"
psycopg2-binary = "^2.9.9"
pytest-asyncio = "^0.23.8"
python-jose = "^3.3.0"
starlette = "^0.38.2"
fastapi-users = "13.0.0"
passlib = "^1.7.4"
greenlet = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]
testpaths = ["tests"]